minimizing debounce duplication
Transferring debounce to utils and calling it from there to reduce duplicates
